Just cleaned up some comments

+++ ForestGenerator/src/NGLScene.cpp
@@ -101,8 +101,7 @@ void NGLScene::buildVAO()
   m_vao=ngl::VAOFactory::createVAO(ngl::simpleIndexVAO,GL_LINES);
   m_vao->bind();
 
-  // set our data for the VAO (when I've sorted this out later, I'll
-  //create the VAO in a different class)
+  // set our data for the VAO
   m_vao->setData(ngl::SimpleIndexVAO::VertexData(
                                         sizeof(ngl::Vec3)*m_LSystem.m_vertices.size(),
                                         m_LSystem.m_vertices[0].m_x,
@@ -111,7 +110,7 @@ void NGLScene::buildVAO()
                                         GL_UNSIGNED_SHORT));
   // data is 12 bytes apart (=sizeof(Vec3))
   m_vao->setVertexAttributePointer(0,3,GL_FLOAT,12,0);
-  m_vao->setNumIndices(size_t(m_LSystem.m_indices.size()));
+  m_vao->setNumIndices(m_LSystem.m_indices.size());
 
   // now unbind
   m_vao->unbind();
